refactor(flow): move navigation here
